Move thematic related routes to separate file
