+++ app/scripts/main.tsx
@@ -1,15 +1,10 @@
 import React, { lazy, Suspense, useEffect } from 'react';
 import { render } from 'react-dom';
 import T from 'prop-types';
-import {
-  BrowserRouter,
-  Navigate,
-  Route,
-  Routes,
-  useLocation,
-  useParams
-} from 'react-router-dom';
+import { BrowserRouter, Route, Routes, useLocation } from 'react-router-dom';
 import { DevseedUiThemeProvider as DsTp } from '@devseed-ui/theme-provider';
+
+import { thematicRoutes } from './thematic-redirect';
 
 import theme, { GlobalStyles } from '$styles/theme';
 import { getAppURL } from '$utils/history';
@@ -30,7 +25,6 @@ const RootDevelopment = lazy(() => import('$components/root-development'));
 const RootDiscoveries = lazy(() => import('$components/root-discoveries'));
 const DataCatalog = lazy(() => import('$components/data-catalog'));
 
-const Home = lazy(() => import('$components/home'));
 // const DiscoveriesHub = lazy(() => import('$components/discoveries/hub'));
 const DiscoveriesSingle = lazy(() => import('$components/discoveries/single'));
 
@@ -70,26 +64,6 @@ function ScrollTop() {
   }, [pathname]);
   return null;
 }
-
-function ThematicAboutRedirect() {
-  const { thematicId } = useParams();
-  return <Navigate replace to={`/${thematicId}`} />;
-}
-
-function ThematicDiscoveryRedirect() {
-  const { discoveryId } = useParams();
-  return <Navigate replace to={`/discoveries/${discoveryId}`} />;
-}
-
-function ThematicDatasetRedirect({ explore = false }: { explore?: boolean }) {
-  const { datasetId } = useParams();
-  let url = `/data-catalog/${datasetId}`;
-  if (explore) {
-    url += '/explore';
-  }
-  return <Navigate replace to={url} />;
-}
-
 // Root component.
 function Root() {
   useScrollbarWidthAsCssVar();
@@ -137,46 +111,8 @@ function Root() {
                   <Route path='/sandbox/*' element={<Sandbox />} />
                 )}
 
-                {/* The following routes are redirect from the legacy thematic areas structure */}
-                <Route path=':thematicId'>
-                  {/* TODO : Redirect to discoveries with filters preset to thematic? */}
-                  {/* TODO : Enable redirection when we want to get rid of thematic area index pages */}
-                  {/* <Route index element={<Navigate replace to='/' />} /> */}
-                  <Route index element={<Home />} />
-
-                  {/* TODO : Redirect to discoveries with filters preset to thematic? */}
-                  <Route
-                    path='discoveries'
-                    element={<Navigate replace to={DISCOVERIES_PATH} />}
-                  />
-                  <Route
-                    path='discoveries/:discoveryId'
-                    element={<ThematicDiscoveryRedirect />}
-                  />
-
-                  {/* TODO : Redirect to data-catalog with filters preset to thematic? */}
-                  <Route
-                    path='datasets'
-                    element={<Navigate replace to='/data-catalog' />}
-                  />
-                  <Route
-                    path='datasets/:datasetId'
-                    element={<ThematicDatasetRedirect />}
-                  />
-                  <Route
-                    path='datasets/:datasetId/explore'
-                    element={<ThematicDatasetRedirect explore />}
-                  />
-                  <Route
-                    path='analysis'
-                    element={<Navigate replace to='/analysis' />}
-                  />
-                  <Route
-                    path='analysis/results'
-                    element={<Navigate replace to='/analysis/results' />}
-                  />
-                  <Route path='about' element={<ThematicAboutRedirect />} />
-                </Route>
+                {/* Legacy: Routes related to thematic areas redirect. */}
+                {thematicRoutes}
 
                 <Route path='*' element={<UhOh />} />
               </Route>
